+++ src/utils/webrtc/models/CallParticipantCollection.js
@@ -19,6 +19,8 @@
  *
  */
 
+import EmitterMixin from '../../EmitterMixin'
+
 import CallParticipantModel from './CallParticipantModel'
 
 /**
@@ -26,48 +28,13 @@ import CallParticipantModel from './CallParticipantModel'
  */
 export default function CallParticipantCollection() {
 
-	this.callParticipantModels = []
+	this._superEmitterMixin()
 
-	this._handlers = []
+	this.callParticipantModels = []
 
 }
 
 CallParticipantCollection.prototype = {
-
-	on(event, handler) {
-		if (!Object.prototype.hasOwnProperty.call(this._handlers, event)) {
-			this._handlers[event] = [handler]
-		} else {
-			this._handlers[event].push(handler)
-		}
-	},
-
-	off(event, handler) {
-		const handlers = this._handlers[event]
-		if (!handlers) {
-			return
-		}
-
-		const index = handlers.indexOf(handler)
-		if (index !== -1) {
-			handlers.splice(index, 1)
-		}
-	},
-
-	_trigger(event, args) {
-		let handlers = this._handlers[event]
-		if (!handlers) {
-			return
-		}
-
-		args.unshift(this)
-
-		handlers = handlers.slice(0)
-		for (let i = 0; i < handlers.length; i++) {
-			const handler = handlers[i]
-			handler.apply(handler, args)
-		}
-	},
 
 	add(options) {
 		const callParticipantModel = new CallParticipantModel(options)
@@ -102,3 +69,5 @@ CallParticipantCollection.prototype = {
 	},
 
 }
+
+EmitterMixin.apply(CallParticipantCollection.prototype)
