+++ src/utils/webrtc/models/LocalCallParticipantModel.js
@@ -19,6 +19,7 @@
  *
  */
 
+import EmitterMixin from '../../EmitterMixin'
 import store from '../../../store/index.js'
 
 import { ConnectionState } from './CallParticipantModel'
@@ -28,6 +29,8 @@ import { ConnectionState } from './CallParticipantModel'
  */
 export default function LocalCallParticipantModel() {
 
+	this._superEmitterMixin()
+
 	this.attributes = {
 		peerId: null,
 		peer: null,
@@ -35,8 +38,6 @@ export default function LocalCallParticipantModel() {
 		guestName: null,
 		connectionState: null,
 	}
-
-	this._handlers = []
 
 	this._handleForcedMuteBound = this._handleForcedMute.bind(this)
 	this._handleExtendedIceConnectionStateChangeBound = this._handleExtendedIceConnectionStateChange.bind(this)
@@ -53,45 +54,6 @@ LocalCallParticipantModel.prototype = {
 		this.attributes[key] = value
 
 		this._trigger('change:' + key, [value])
-	},
-
-	on(event, handler) {
-		if (!Object.prototype.hasOwnProperty.call(this._handlers, event)) {
-			this._handlers[event] = [handler]
-		} else {
-			this._handlers[event].push(handler)
-		}
-	},
-
-	off(event, handler) {
-		const handlers = this._handlers[event]
-		if (!handlers) {
-			return
-		}
-
-		const index = handlers.indexOf(handler)
-		if (index !== -1) {
-			handlers.splice(index, 1)
-		}
-	},
-
-	_trigger(event, args) {
-		let handlers = this._handlers[event]
-		if (!handlers) {
-			return
-		}
-
-		if (!args) {
-			args = []
-		}
-
-		args.unshift(this)
-
-		handlers = handlers.slice(0)
-		for (let i = 0; i < handlers.length; i++) {
-			const handler = handlers[i]
-			handler.apply(handler, args)
-		}
 	},
 
 	setWebRtc(webRtc) {
@@ -193,3 +155,5 @@ LocalCallParticipantModel.prototype = {
 	},
 
 }
+
+EmitterMixin.apply(LocalCallParticipantModel.prototype)
