+++ src/utils/webrtc/models/LocalMediaModel.js
@@ -19,12 +19,15 @@
  *
  */
 
+import EmitterMixin from '../../EmitterMixin'
 import store from '../../../store/index.js'
 
 /**
  *
  */
 export default function LocalMediaModel() {
+
+	this._superEmitterMixin()
 
 	this.attributes = {
 		localStreamRequestVideoError: null,
@@ -41,8 +44,6 @@ export default function LocalMediaModel() {
 		token: '',
 		raisedHand: false,
 	}
-
-	this._handlers = []
 
 	this._handleLocalStreamRequestedBound = this._handleLocalStreamRequested.bind(this)
 	this._handleLocalStreamBound = this._handleLocalStream.bind(this)
@@ -74,41 +75,6 @@ LocalMediaModel.prototype = {
 		this.attributes[key] = value
 
 		this._trigger('change:' + key, [value])
-	},
-
-	on(event, handler) {
-		if (!Object.prototype.hasOwnProperty.call(this._handlers, event)) {
-			this._handlers[event] = [handler]
-		} else {
-			this._handlers[event].push(handler)
-		}
-	},
-
-	off(event, handler) {
-		const handlers = this._handlers[event]
-		if (!handlers) {
-			return
-		}
-
-		const index = handlers.indexOf(handler)
-		if (index !== -1) {
-			handlers.splice(index, 1)
-		}
-	},
-
-	_trigger(event, args) {
-		let handlers = this._handlers[event]
-		if (!handlers) {
-			return
-		}
-
-		args.unshift(this)
-
-		handlers = handlers.slice(0)
-		for (let i = 0; i < handlers.length; i++) {
-			const handler = handlers[i]
-			handler.apply(handler, args)
-		}
 	},
 
 	getWebRtc() {
@@ -462,3 +428,5 @@ LocalMediaModel.prototype = {
 	},
 
 }
+
+EmitterMixin.apply(LocalMediaModel.prototype)
